+++ src/dfx/src/lib/builders/custom.rs
@@ -113,7 +113,6 @@ impl CanisterBuilder for CustomBuilder {
             dependencies,
         } = CustomBuilderExtra::try_from(info, pool, no_deps)?;
 
-        let canister_id = info.get_canister_id().unwrap();
         let vars = super::get_and_write_environment_variables(
             info,
             &config.network_name,
@@ -135,7 +134,6 @@ impl CanisterBuilder for CustomBuilder {
         }
 
         Ok(BuildOutput {
-            canister_id,
             wasm: WasmBuildOutput::File(wasm),
             idl: IdlBuildOutput::File(info.get_output_idl_path().to_path_buf()),
         })